# optimize gc sweep